fork: fix incorrect debug message when SO_LINGER is enabled
Since 9436e7b (fork: add an option to avoid SO_LINGER (-1) or a
different timeout, 2023-11-05), a debug message was printed if
SO_LINGER was disabled (ex: when using `--linger=-1`).

The last minute logic used was incorrect and would result in a
message incorrectly stating that SO_LINGER was disabled even if
valid timeouts (>=0) were provided and accepted (an error to set
SO_LINGER itself would result in a fatal error instead).