client.pl: add --*timeout for connect/read values
As an alternative to changing the code, allow the use of --connect or
--read options to indicate how many seconds to use on each case.

Note that to simplify the interface, the read timeout provided will be
used for all commands, and therefore if using a high latency/low bandwith
connection might might need to make sure the "big" command can complete
normally with the default (5 seconds) or provided value.

While at it, move the variables (including the one used to describe the
recovery mode for non interactive, at connect, commands) to the beginning
of the file.