client: implement cork command
Allow setting TCP_CORK (mainly useful in Linux) to allow bundling
multiple operations, so the next command (using ping in the following
example), can execute with a half closed socket by doing:

  passive
  cork
  shut 1
  ping

Make sure that known closed connections (or half closed in this case)
are identified and terminate the loop, to ensure a clean connection
close.

While at it, do some minor cleanup in the module and export list.