fork: implement --cork option
Allow enabling TCP_CORK (if supported or an "equivalent" is available)
to delay sending the written response until the socket closes, so it
will be hopefully sent with a combined packet.

Since there is a risk of lost data (specially when not doing a explicit
close) add a warning at startup.

While at it, reorder comand line options and fix a documentation typo.