client.pl: improve cork support for BSD and macOS
In a previous commit, cork mode was introduced using the TCP_CORK
option, which is not supported at least in BSD.

Add a runtime heuristic based on the OS name to use TCP_NOPUSH as
an alternative.

Note, there is at least one known bug with the "detection" code,
and that it leaves the variable undefined for other systems.