client.pl: avoid ascii modifier in regex for compatibility
Perl older than 5.12 will be confused by it, so instead make
it the default through the locale module.

Note that all commands are still ASCII only, so anythin non
7bit clean bytes will be sent AS-IS.

Alternatiely, changing the regex to use the literal `[0-9]`
for digits might be possible, but punted for now to avoid
making the code more complicated thn necessary.

While at it, move the core modules to the top of the script
so the config variables are more easily identifiable.